+++ cmem_plugin_kafka/utils.py
@@ -1,12 +1,9 @@
 """Kafka utils modules"""
-import gzip
 import json
 import re
 from typing import Dict, Any, Iterator, Optional
 from urllib.parse import urlparse
 
-import lz4.frame as lz4frame
-import zstandard
 from cmem.cmempy.config import get_cmem_base_uri
 from cmem.cmempy.workspace.projects.resources.resource import get_resource_response
 from cmem.cmempy.workspace.search import list_items
@@ -26,7 +23,6 @@ from cmem_plugin_base.dataintegration.utils import (
 from confluent_kafka import Producer, Consumer, KafkaException, KafkaError
 from confluent_kafka.admin import AdminClient, TopicMetadata, ClusterMetadata
 from defusedxml import ElementTree
-from snappy import snappy
 
 from cmem_plugin_kafka.constants import KAFKA_TIMEOUT
 
@@ -66,45 +62,24 @@ class KafkaProducer:
 
     def process(self, message: KafkaMessage):
         """Produce message to topic."""
-        self._no_of_success_messages += 1
+        # self._no_of_success_messages += 1
         headers = message.headers if message.headers else {}
-        if self.compression_type != 'none' :
-            headers['compression.type'] = self.compression_type
         self._producer.produce(
             self._topic,
-            value=self.compress(message.value),
+            value=message.value.encode('utf-8'),
             key=message.key,
-            headers=headers
+            headers=headers,
+            on_delivery=self.on_delivery
         )
 
-    def compress(self, value: str):
+    def on_delivery(self, err, msg):
         """
-        Compresses the given value based on the configured compression type.
-
-        Args:
-            value (bytes): The value to compress.
-
-        Returns:
-            bytes: The compressed value.
-
-        Raises:
-            ValueError: If an unsupported compression type is provided.
+        Callback method executed after a message is delivered to the Kafka broker.
         """
-        _ = value.encode('utf-8')
-        if self.compression_type == 'none':
-            return _
-        if self.compression_type == 'gzip':
-            return gzip.compress(_)
-        if self.compression_type == 'snappy':
-            return snappy.compress(_)
-        if self.compression_type == 'zstd':
-            compressor = zstandard.ZstdCompressor(level=3)
-            compressed_data = compressor.compress(_)
-            return compressed_data
-        if self.compression_type == 'lz4':
-            compressed_data = lz4frame.compress(_)
-            return compressed_data
-        raise ValueError(f'Unsupported compression type: {self.compression_type}')
+        _ = msg
+        if err:
+            raise KafkaException(err)
+        self._no_of_success_messages += 1
 
     def poll(self, timeout):
         """Polls the producer for events and calls the corresponding callbacks"""
@@ -137,6 +112,7 @@ class KafkaConsumer:
         self._log = log
         self._no_of_success_messages = 0
         self._first_message: Optional[KafkaMessage] = None
+        self.fetch_limit = -1
 
     def get_success_messages_count(self) -> int:
         """Return count of the successful messages"""
@@ -173,6 +149,10 @@ class KafkaConsumer:
     def poll(self) -> Iterator[KafkaMessage]:
         """Polls the consumer for events and calls the corresponding callbacks"""
         while True:
+            if 0 <= self.fetch_limit == self._no_of_success_messages:
+                self._log.info("Message fetch fetch_limit reached")
+                break
+
             msg = self._consumer.poll(timeout=KAFKA_TIMEOUT)
             if msg is None:
                 self._log.info("Messages are empty")
@@ -182,29 +162,11 @@ class KafkaConsumer:
                 raise KafkaException(msg.error())
 
             self._no_of_success_messages += 1
-            headers = dict(msg.headers()) if msg.headers() else {}
-            if headers and 'compression.type' in headers:
-                compression_type = headers['compression.type']
-                if compression_type == b'gzip':
-                    decompressed_message = gzip.decompress(msg.value())
-                elif compression_type == b'snappy':
-                    decompressed_message = snappy.decompress(msg.value())
-                elif compression_type == b'zstd':
-                    decompressor = zstandard.ZstdDecompressor()
-                    decompressed_message = decompressor.decompress(msg.value())
-                elif compression_type == b'lz4':
-                    decompressed_message = lz4frame.decompress(msg.value())
-                else:
-                    raise ValueError(
-                        f'Unsupported compression codec: {compression_type}'
-                    )
-            else:
-                decompressed_message = msg.value()
 
             kafka_message = KafkaMessage(
                 key=msg.key().decode("utf-8") if msg.key() else "",
                 headers=msg.headers(),
-                value=decompressed_message.decode("utf-8"),
+                value=msg.value().decode("utf-8"),
             )
 
             if not self._first_message: